+++ Makefile
@@ -27,34 +27,13 @@ setup: ## Set up local environment for Python development on pipelines
 	cd pipelines && \
 	pipenv install --dev
 
-setup-components: ## Setup unit tests for a component group
-	@cd pipelines/pipelines/kfp_components/${GROUP} && \
-	pipenv install --dev
-
-setup-all-components: ## Setup unit tests for all pipeline components
-	@set -e && \
-	for component_group in pipelines/pipelines/kfp_components/*/ ; do \
-		$(MAKE) setup-components GROUP=$$(basename $$component_group) ; \
-	done
-
 test-trigger: ## Runs unit tests for the pipeline trigger code
 	@cd pipelines && \
 	pipenv run python -m pytest tests/trigger
 
 compile-pipeline: ## Compile the pipeline to training.json or prediction.json. Must specify pipeline=<training|prediction>
-	@cd pipelines && \
+	@cd pipelines/src && \
 	pipenv run python -m pipelines.${PIPELINE_TEMPLATE}.${pipeline}.pipeline
-
-test-components: ## Run unit tests for a component group
-	@cd pipelines/pipelines/kfp_components/${GROUP} && \
-	pipenv run python -m pytest
-
-test-all-components: ## Run unit tests for all pipeline components
-	@set -e && \
-	for component_group in pipelines/pipelines/kfp_components/*/ ; do \
-		echo "Running unit tests for components under $$component_group" && \
-		$(MAKE) test-components GROUP=$$(basename $$component_group) ; \
-	done
 
 sync-assets: ## Sync assets folder to GCS. Must specify pipeline=<training|prediction>
 	if [ -d "./pipelines/pipelines/${PIPELINE_TEMPLATE}/$(pipeline)/assets/" ] ; then \
@@ -64,14 +43,14 @@ sync-assets: ## Sync assets folder to GCS. Must specify pipeline=<training|predi
 run: ## Compile pipeline, copy assets to GCS, and run pipeline in sandbox environment. Must specify pipeline=<training|prediction>. Optionally specify enable_pipeline_caching=<true|false> (defaults to default Vertex caching behaviour)
 	@ $(MAKE) compile-pipeline && \
 	$(MAKE) sync-assets && \
-	cd pipelines && \
-	pipenv run python -m trigger.main --template_path=./$(pipeline).json --enable_caching=$(enable_pipeline_caching)
+	cd pipelines/src && \
+	pipenv run python -m pipelines.trigger --template_path=./$(pipeline).json --enable_caching=$(enable_pipeline_caching)
 
 e2e-tests: ## Compile pipeline, copy assets to GCS, and perform end-to-end (E2E) pipeline tests. Must specify pipeline=<training|prediction>. Optionally specify enable_pipeline_caching=<true|false> (defaults to default Vertex caching behaviour)
 	@ $(MAKE) compile-pipeline && \
 	$(MAKE) sync-assets && \
 	cd pipelines && \
-	pipenv run python -m pytest --log-cli-level=INFO tests/${PIPELINE_TEMPLATE}/$(pipeline) --enable_caching=$(enable_pipeline_caching)
+	pipenv run python pytest --log-cli-level=INFO tests/${PIPELINE_TEMPLATE}/$(pipeline) --enable_caching=$(enable_pipeline_caching)
 
 env ?= dev
 deploy-infra: ## Deploy the Terraform infrastructure to your project. Requires VERTEX_PROJECT_ID and VERTEX_LOCATION env variables to be set in env.sh. Optionally specify env=<dev|test|prod> (default = dev)
