+++ aiplatform-components/Pipfile
@@ -4,11 +4,11 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
-kfp = "==1.8.9"
-google-cloud-pipeline-components = ">=1.0.0,<2.0.0"
-google-cloud-aiplatform = ">=1.24.1,<2.0.0"
+kfp = "==1.8.21"
 
 [dev-packages]
+google-cloud-aiplatform = ">= 1.24.1,<2.0.0"
+google-cloud-pipeline-components = ">= 1.0.0,<2.0.0"
 pytest = ">=6.2.4,<7.0.0"
 pre-commit = ">=2.14.1,<3.0.0"
