+++ aiplatform-components/src/aiplatform_components/import_model_evaluation.py
@@ -19,7 +19,7 @@ from typing import NamedTuple
 @component(
     base_image="python:3.7",
     packages_to_install=[
-        "google-cloud-aiplatform>=1.24.1",
+        "google-cloud-aiplatform==1.24.1",
     ],
 )
 def import_model_evaluation(
