+++ bigquery-components/.python-version
