+++ bigquery-components/Pipfile
@@ -4,7 +4,7 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
-kfp = "==1.8.9"
+kfp = "==1.8.21"
 
 [dev-packages]
 pytest = ">=6.2.4,<7.0.0"
