+++ pipelines/Pipfile
@@ -4,10 +4,12 @@ verify_ssl = true
 name = "pypi"
 
 [packages]
-kfp = "==1.8.9"
+kfp = "==1.8.21"
 google-cloud-pipeline-components = ">=1.0.0,<2.0.0"
 google-cloud-aiplatform = ">=1.24.1,<2.0.0"
 Jinja2 = ">=3.0.1,<4.0.0"
+aiplatform-components = {editable = true, path = "./../aiplatform-components"}
+bigquery-components = {editable = true, path = "./../bigquery-components"}
 
 [dev-packages]
 pytest = ">=6.2.4,<7.0.0"
