+++ pipelines/src/pipelines/tensorflow/prediction/pipeline.py
@@ -18,12 +18,9 @@ import pathlib
 
 from kfp.v2 import compiler, dsl
 
-from pipelines import generate_query
-from pipelines.kfp_components import (
-    lookup_model,
-    bq_query_to_table,
-    model_batch_predict,
-)
+from src.pipelines import generate_query
+from bigquery_components import bq_query_to_table
+from aiplatform_components import lookup_model, model_batch_predict
 
 
 @dsl.pipeline(name="tensorflow-prediction-pipeline")
