+++ pipelines/src/pipelines/tensorflow/prediction/queries/ingest.sql
