+++ pipelines/src/pipelines/tensorflow/training/assets/train_tf_model.py
