+++ pipelines/src/pipelines/tensorflow/training/queries/engineer_features.sql
