+++ pipelines/src/pipelines/tensorflow/training/queries/ingest.sql
