+++ pipelines/src/pipelines/tensorflow/training/queries/sample.sql
