+++ pipelines/src/pipelines/trigger/__main__.py
@@ -1,0 +1,5 @@
+from .main import sandbox_run
+
+
+if __name__ == "__main__":
+    sandbox_run()
