+++ pipelines/src/pipelines/trigger/requirements.txt
