+++ pipelines/src/pipelines/xgboost/training/queries/engineer_features.sql
