+++ pipelines/src/pipelines/xgboost/training/queries/ingest.sql
