+++ pipelines/src/pipelines/xgboost/training/queries/sample.sql
